Remove Path annotation, cleanup
