+++ src/main/java/gg/projecteden/jayce/commands/IssuesAppCommand.java
@@ -9,7 +9,6 @@ import gg.projecteden.jayce.commands.common.AppCommand;
 import gg.projecteden.jayce.commands.common.AppCommandEvent;
 import gg.projecteden.jayce.commands.common.annotations.Choices;
 import gg.projecteden.jayce.commands.common.annotations.Desc;
-import gg.projecteden.jayce.commands.common.annotations.Path;
 import gg.projecteden.jayce.commands.common.annotations.Role;
 import gg.projecteden.jayce.config.Config;
 import gg.projecteden.jayce.github.Issues.RepoIssueContext;
@@ -26,7 +25,7 @@ import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
 
-import static gg.projecteden.jayce.commands.common.AppCommandRegistration.supplyChoices;
+import static gg.projecteden.jayce.commands.common.AppCommandRegistry.supplyChoices;
 import static gg.projecteden.utils.StringUtils.ellipsis;
 import static java.util.Objects.requireNonNull;
 
@@ -41,7 +40,6 @@ public class IssuesAppCommand extends AppCommand {
 	}
 
 	@Desc("Create an issue")
-	@Path("create <title> <body>")
 	void create(String title, String body) {
 		issues.create(issues.of(member(), title, body).build()).thenAccept(result -> {
 			if (!isWebhookChannel())
@@ -50,61 +48,51 @@ public class IssuesAppCommand extends AppCommand {
 	}
 
 	@Desc("Open an existing issue")
-	@Path("open <id>")
 	void open(@Desc("issue number") int id) {
 		issues.open(id).thenRun(this::thumbsup);
 	}
 
 	@Desc("Close an issue")
-	@Path("close <id>")
 	void close(@Desc("issue number") int id) {
 		issues.close(id).thenRun(this::thumbsup);
 	}
 
 	@Desc("Add a user to an issue's assignees")
-	@Path("assign <id> <user>")
 	void assign(@Desc("issue number") int id, Member user) {
 		issues.assign(id, user).thenRun(this::thumbsup);
 	}
 
 	@Desc("Remove a user from an issue's assignees")
-	@Path("unassign <id> <user>")
 	void unassign(@Desc("issue number") int id, Member user) {
 		issues.unassign(id, user).thenRun(this::thumbsup);
 	}
 
 	@Desc("Edit an issue's title")
-	@Path("edit title <id> <text>")
 	void edit_title(@Desc("issue number") int id, String text) {
 		issues.edit(id, issue -> issue.withTitle(text)).thenRun(this::thumbsup);
 	}
 
 	@Desc("Edit an issue's body")
-	@Path("edit body <id> <text>")
 	void edit_body(@Desc("issue number") int id, String text) {
 		issues.edit(id, issue -> issue.withBody(Optional.of(text))).thenRun(this::thumbsup);
 	}
 
 	@Desc("Comment on an issue")
-	@Path("comment <id> <text>")
 	void comment(@Desc("issue number") int id, String text) {
 		issues.comment(id, "**" + name() + "**: " + text).thenRun(this::thumbsup);
 	}
 
 	@Desc("Add a label to an issue")
-	@Path("labels add <id> <label>")
 	void labels_add(@Desc("issue number") int id, @Choices(Label.class) String label) {
 		issues.addLabels(id, List.of(label)).thenRun(this::thumbsup);
 	}
 
 	@Desc("Remove a label from an issue")
-	@Path("labels remove <id> <label>")
 	void labels_remove(@Desc("issue number") int id, @Choices(Label.class) String label) {
 		issues.removeLabels(id, List.of(label)).thenRun(this::thumbsup);
 	}
 
 	@Desc("Search existing issues")
-	@Path("search <query>")
 	void search(String query) {
 		issues.search(query).thenAccept(items -> {
 			if (Utils.isNullOrEmpty(items))
@@ -127,7 +115,6 @@ public class IssuesAppCommand extends AppCommand {
 	}
 
 	@Desc("Close all issues in a repository")
-	@Path("closeAll <repo>")
 	void closeAll(String repo) {
 		if (Jayce.get().getEnv() != Env.DEV)
 			throw new EdenException("Development environment only command");
