+++ src/main/java/gg/projecteden/jayce/commands/common/annotations/NotRequired.java
@@ -5,9 +5,8 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-@Target(ElementType.METHOD)
+@Target(ElementType.PARAMETER)
 @Retention(RetentionPolicy.RUNTIME)
-public @interface Path {
-	String value() default "";
+public @interface NotRequired {
 
 }
